+++ tests/integration_tests/utils.py
@@ -51,6 +51,7 @@ TEST_CONTRACTS = {
     "GasConsumerTryCatch": "GasConsumerTryCatch.sol",
     "DelegationTarget": "DelegationTarget.sol",
     "MaliciousDeployer": "MaliciousDeployer.sol",
+    "SelfDestructExploitFactory": "SelfDestructExploit.sol",
 }